Did a lot of things omg

+++ app/src/main/java/com/example/HAHA/Data/PaymentData.kt
@@ -3,6 +3,5 @@ package com.example.HAHA.Data
 data class PaymentData(
     val item: String,
     val desc: String,
-    val amount: Int
-
+    val amount: Double
 )
